pyproject.toml
[build-system]
requires = ["setuptools ~= 69.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "investir"
version = "1.7"
dependencies = [
  "click ~= 8.1.0",
  "py-moneyed ~= 3.0",
  "platformdirs ~= 4.2.0",
  "prettytable ~= 3.10.0",
  "python-dateutil ~= 2.9.0",
  "PyYAML ~= 6.0.0",
  "typer ~= 0.20.0",
  "yfinance ~= 1.0",
  "pip>=26.0.1",
]
optional-dependencies.lint = [
  "mypy ~= 1.18.0",
  "ruff ~= 0.14.0",
]
optional-dependencies.test = [
  "pandas == 2.2.3",
  "pytest ~= 8.1.0",
  "pytest-cov ~= 5.0.0",
  "pytest-mock ~= 3.14.0"
]
requires-python = ">=3.10"
authors = [
  {name = "Tiago Gomes", email = "tiago.avv+oss@gmail.com"},
]
maintainers = [
  {name = "Tiago Gomes", email = "tiago.avv+oss@gmail.com"}
]
description = "Calculate UK capital gains tax for Freetrade and Trading212"
readme = "README.md"
license = {text = "MIT License"}
keywords = ["HMRC", "Shares", "Stock", "Tax", "Capital Gains", "Freetrade", "Trading212"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Office/Business",
  "Topic :: Office/Business :: Financial",
  "Topic :: Office/Business :: Financial :: Accounting",
  "Topic :: Office/Business :: Financial :: Investment",
  "Typing :: Typed"
]

[project.scripts]
investir = "investir.cli:main"

[project.urls]
Homepage = "https://github.com/tacgomes/investir"
Repository = "https://github.com/tacgomes/investir"
"Bug Tracker" = "https://github.com/tacgomes/investir/issues"

[[tool.mypy.overrides]]
module = "yfinance"
ignore_missing_imports = true

[tool.ruff]
lint.select = [
    "F",   # pyflakes
    "W",   # pycodestyle warning
    "E",   # pycodestyle error
    "B",   # flake8 bugbear plugin
    "PL",  # Pylint
    "I",   # isort
    "RUF", # Ruff
]
lint.ignore = [
  "PLC0415",  # Import should be at the top-level of a file
  "PLR2004",  # Magic value used in comparison.
  "PLR0913",  # Too many arguments in function definition.
]

[tool.coverage.run]
source_pkgs = ["investir"]

[tool.coverage.report]
show_missing = true
exclude_also = [
    "class \\w+\\(Protocol\\):",
    "def main()",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
addopts = "--cov --no-header"
markers = [
    """network: mark a test that requires an Internet \
    connection (deselect with '-m "not network"')"""
]

[dependency-groups]
dev = [
    "pip>=26.0.1",
]
